Move virtual garden rules
